requirements.txt
numpy==1.26.4
blosc2==3.6.1
cython
streamlit==1.26.0
langchain==0.0.352
faiss-cpu==1.11.0.post1
pandas
PyPDF2
chromadb
groq
openai
scikit-learn
python-dotenv
sentence-transformers





